stub: adds initial work on version solving
Admittedly, this is **very** unfinished (read: completely unusable)
but it does lay out the main ideas I'll be using here.

I'm going to be focusing on Wave function collapse generation for a bit
as it makes me feel better than working on this.

I also might be doing some refactors based on
re-examining commonalities to constraint solving problems
as I suspect the current design I have may not be the best...